Kovalchuk/lab07/requirements.txt
pygame==1.9.2b8
PyOpenGL==3.1.0
